+++ ecosystem.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+	apps: [{
+		script: './src/index.js',
+		watch: 'true',
+		ignore_watch: ['node_modules', 'src/databases'],
+	}],
+};