+++ src/events/guildCreate.js
@@ -1,5 +1,6 @@
 const { createDatabase } = require('../helpers/dbModel');
 const { Permissions } = require('discord.js');
+const { setPerms } = require('../helpers/setCommandPerm');
 
 module.exports = {
 	name: 'guildCreate',
@@ -26,27 +27,11 @@ module.exports = {
 
 		// Add admin commands role perm
 		guild.roles.cache.filter(role => role.permissions.has(Permissions.FLAGS.ADMINISTRATOR) && !role.managed).each(adminRole => {
-			// guild.commands.fetch()                       // Guild commands
-			guild.client.application?.commands.fetch() // Global commands
-				.then(async commands => {
-					for (const name of adminCommands) {
-						const foundCommand = await commands.find(command => command.name === name);
-						const permission = [{ id: adminRole.id, type: 'ROLE', permission: true }];
-						await foundCommand.permissions.add({ guild: guild.id, permissions: permission });
-					}
-				}).catch(e => console.error(e.toString())); // Unable to fetch guild commands.
+			setPerms(adminRole, adminCommands, true);
 		});
 		// Add manage emojis commands role perm
 		guild.roles.cache.filter(role => role.permissions.has(Permissions.FLAGS.MANAGE_EMOJIS_AND_STICKERS) && !role.managed).each(manageEmojisRole => {
-			// guild.commands.fetch()                       // Guild commands
-			guild.client.application?.commands.fetch() // Global commands
-				.then(async commands => {
-					for (const name of manageEmojisCommands) {
-						const foundCommand = await commands.find(command => command.name === name);
-						const permission = [{ id: manageEmojisRole.id, type: 'ROLE', permission: true }];
-						await foundCommand.permissions.add({ guild: guild.id, permissions: permission });
-					}
-				}).catch(e => console.error(e.toString())); // Unable to fetch guild commands.
+			setPerms(manageEmojisRole, manageEmojisCommands, true);
 		});
 	},
 };