+++ src/helpers/deployCommands.js
@@ -5,10 +5,10 @@ const { Routes } = require('discord-api-types/v9');
 const { clientId, guildId, token } = require('../../config.json');
 
 const commands = [];
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
+const commandFiles = fs.readdirSync('./src/commands').filter(file => file.endsWith('.js'));
 
 for (const file of commandFiles) {
-	const command = require(`./src/commands/${file}`);
+	const command = require(`../commands/${file}`);
 	commands.push(command.data.toJSON());
 }
 
@@ -17,7 +17,7 @@ const rest = new REST({ version: '9' }).setToken(token);
 (async () => {
 	try {
 		await rest.put(
-			// Routes.applicationGuildCommands(clientId, guildId),   // <-- Guild version
+			// Routes.applicationGuildCommands(clientId, guildId), // <-- Guild version
 			Routes.applicationCommands(clientId), // <-- Global version
 			{ body: commands },
 		);